requirements.txt
tinytuya
requests
python-telegram-bot
configparser 
